% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMatrix.R
\name{initMatrix}
\alias{initMatrix}
\title{Initailised a matrix of 2D}
\usage{
initMatrix(size, debug = FALSE)
}
\arguments{
\item{size}{\code{numeric} of array, default based on L2 Cache}

\item{debug}{\code{boolean} set to show messages}
}
\description{
Initailised a matrix of 2D
}
