% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparallelcuda-package.R
\docType{package}
\name{rparallelcuda-package}
\alias{rparallelcuda}
\alias{rparallelcuda-package}
\title{rparallelcuda: R Parallel Cuda}
\description{
PoC of parallelism handling oneAPI and CUDA.
}
\author{
\strong{Maintainer}: Renato Samperio \email{renato.samperio@zurich.com}

}
\keyword{internal}
