% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cuda.R
\name{run_cuda}
\alias{run_cuda}
\title{Runs CUDA kernel}
\usage{
run_cuda(size = 500, debug = FALSE)
}
\arguments{
\item{size}{\code{numeric} size of arrays}

\item{debug}{\code{bool} show debug messages}
}
\description{
Runs CUDA kernel
}
