% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_matrixes_cuda}
\alias{sum_matrixes_cuda}
\title{Function to sum matrixes using TBB and CUDA}
\usage{
sum_matrixes_cuda(A, B, C, N, block1 = 16L, block2 = 16L)
}
\arguments{
\item{A}{\code{matrix} array of input numbers}

\item{B}{\code{matrix} array of input numbers}

\item{C}{\code{matrix} array of input numbers}

\item{N}{\code{numeric} of array, default based on L2 Cache}

\item{block1}{\code{numeric} size of block 1}

\item{block2}{\code{numeric} size of block 2}
}
\description{
Function to sum matrixes using TBB and CUDA
}
