% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_matrixes_tbb_graph}
\alias{sum_matrixes_tbb_graph}
\title{TBB parallelism function using a graph of nodes}
\usage{
sum_matrixes_tbb_graph(A, B, C, N)
}
\arguments{
\item{A}{\code{vector} array of input numbers}

\item{B}{\code{vector} array of input numbers}

\item{C}{\code{vector} array of input numbers}

\item{N}{\code{numeric} of array, default based on L2 Cache}
}
\description{
TBB parallelism function using a graph of nodes
}
