% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_matrixes_tbb_parallel}
\alias{sum_matrixes_tbb_parallel}
\title{TBB using parallel for}
\usage{
sum_matrixes_tbb_parallel(A, B, C, N)
}
\arguments{
\item{A}{\code{vector} array of input numbers}

\item{B}{\code{vector} array of input numbers}

\item{C}{\code{vector} array of input numbers}

\item{N}{\code{numeric} of array, default based on L2 Cache}
}
\description{
TBB using parallel for
}
