% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_matrixes_tbb_cuda.R
\name{sumMatrixesTbbCuda}
\alias{sumMatrixesTbbCuda}
\title{Executes a summation with TBB and CUDA}
\usage{
sumMatrixesTbbCuda(
  A = NULL,
  B = NULL,
  C = NULL,
  size = 496,
  debug = FALSE,
  block1 = 16,
  block2 = 16
)
}
\arguments{
\item{A}{\code{vector} array of numbers}

\item{B}{\code{vector} array of numbers}

\item{C}{\code{vector} array of numbers}

\item{size}{\code{numeric} of array, default based on L2 Cache}

\item{debug}{\code{boolean} set to show messages}

\item{block1}{\code{numeric} size of block 1}

\item{block2}{\code{numeric} size of block 2}
}
\description{
Executes a summation with TBB and CUDA
}
