% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_matrixes_tbb_parallel.R
\name{sumMatrixesTbbParallel}
\alias{sumMatrixesTbbParallel}
\title{Executes a summation with TBB and CUDA}
\usage{
sumMatrixesTbbParallel(A = NULL, B = NULL, C = NULL, size = 496, debug = FALSE)
}
\arguments{
\item{A}{\code{vector} array of numbers}

\item{B}{\code{vector} array of numbers}

\item{C}{\code{vector} array of numbers}

\item{size}{\code{numeric} of array, default based on L2 Cache}

\item{debug}{\code{boolean} set to show messages}
}
\description{
Executes a summation with TBB and CUDA
}
