% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_matrix_cuda.R
\name{test_cuda_sum}
\alias{test_cuda_sum}
\title{Run parallel benchmark}
\usage{
test_cuda_sum(
  A = NULL,
  B = NULL,
  C = NULL,
  times = 1,
  size = 496,
  debug = FALSE,
  with_gc = FALSE,
  functor = sumMatrixesTbbCuda
)
}
\arguments{
\item{A}{\code{vector} array of numbers}

\item{B}{\code{vector} array of numbers}

\item{C}{\code{vector} array of numbers}

\item{times}{\code{numeric} how many times it shoud run}

\item{size}{\code{numeric} size of vector if not provided}

\item{debug}{\code{boolean} set to show messages}

\item{with_gc}{\code{boolean} set to include garbage collection}

\item{functor}{\code{function} call for CUDA kernel}
}
\description{
Run parallel benchmark
}
