% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_matrix_sum.R
\name{test_parallel_sum}
\alias{test_parallel_sum}
\title{Run parallel benchmark}
\usage{
test_parallel_sum(
  A = NULL,
  B = NULL,
  C = NULL,
  times = 1,
  size = 496,
  debug = FALSE
)
}
\arguments{
\item{A}{\code{vector} array of numbers}

\item{B}{\code{vector} array of numbers}

\item{C}{\code{vector} array of numbers}

\item{times}{\code{numeric} how many times it shoud run}

\item{size}{\code{numeric} size of vector if not provided}

\item{debug}{\code{boolean} set to show messages}
}
\description{
Run parallel benchmark
}
